*Conveyor and *Generator classes renamed; cleanup
