+++ tests/Conveyor/Stub/FakeGenerator.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Yiisoft\Yii\Cycle\Tests\Generator\Stub;
+namespace Yiisoft\Yii\Cycle\Tests\Conveyor\Stub;
 
 use Cycle\Schema\GeneratorInterface;
 use Cycle\Schema\Registry;
